paynow/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
paynow
requests
pydantic
httpx
mangum>=0.17.0
slowapi
